Condensed install process and updated README
The installation process for the linuxize process has been massively
simplified by this update. All of the steps of installing config files
(which was previously a massive portion of the install process) has now
been automated and will occur as part of the scoop install process.
Documentation has also been added to the readme which explains how to
use the functions given to the client during the install process. The
documentation now also gives some of the benefits incurred during the
transformation process.