Use official version of FiraCode
The version of FiraCode Nerd Font now installed by the scoop-install
script will be from the official FiraCode git repository rather than one
from mine. This will ensure that the version installed will always be
the updated version. This also allows the repository to purge the /bin
folder and with it the FiraCode ttf file, by far the largest file in the
repostitory.