Added CSS plugin to init.vim
The list of plugins in init.vim has been ammended to use the css colors
vim plugin. This plugin automatically highlights color codes and names
in css files (this works in html files too). The installation of all
plugins is done automatically in the scoop-install script so this will
not break anything in the installation process.