Added groff documentation to README
Documentation on the groff document compiling features has been added to
the readme for this git repository. This description contains
instructions to install groff, which is not installed by default by the
scoop-install script. The rationale for this has also been added. The
description also contains some useful manpages that are helpful to learn
groff.